Updated message entities and added message entities' javadoc Created constructor for abstract Message
